+++ src/systems/target.h
@@ -11,10 +11,10 @@
 class TargetSystem : public System {
  public:
     TargetSystem() : m_sprite{g_textures.load("target.png")} {}
-    void update(double) override {
+    void update(uint64_t) override {
         if (rand() % 100 == 0) {
             Entity newEntity = g_ecs.createEntity();
-            g_ecs.addComponents<CSprite, CTranslation, CVelocity, CTargetTag>(newEntity, CSprite{m_sprite, 50, 50}, CTranslation{(float)100 * (rand() % 4 + 1), 500}, CVelocity{0, -300}, CTargetTag{});
+            g_ecs.addComponents<CSprite, CTranslation, CVelocity, CTargetTag>(newEntity, CSprite{m_sprite, 50, 50}, CTranslation{100.f * (rand() % 4 + 1), 500}, CVelocity{0, -300}, CTargetTag{});
         }
         for (Entity e : m_entities) {
             CTranslation &translation = g_ecs.getComponent<CTranslation>(e);
